refactor: extract useCarouselScroll hook and fix ScrollArrow button type
- Extract scroll state/effect logic into useCarouselScroll hook shared by
  PlaylistCarousel and ArtistDetail, eliminating code duplication
- Fix useEffect dependency to use stable albumsData?.items instead of
  the albums array recreated on every render
- Add type="button" to ScrollArrow button to prevent accidental form submit
