+++ src/hooks/useCarouselScroll.ts
@@ -1,0 +1,34 @@
+import { useRef, useState, useEffect } from 'react';
+
+const CARD_WIDTH = 168;
+
+export function useCarouselScroll(dependency: unknown) {
+  const scrollRef = useRef<HTMLDivElement>(null);
+  const [canScrollLeft, setCanScrollLeft] = useState(false);
+  const [canScrollRight, setCanScrollRight] = useState(false);
+
+  useEffect(() => {
+    const el = scrollRef.current;
+    if (!el) return;
+
+    const updateArrows = () => {
+      setCanScrollLeft(el.scrollLeft > 0);
+      setCanScrollRight(el.scrollLeft + el.clientWidth < el.scrollWidth - 1);
+    };
+
+    updateArrows();
+    el.addEventListener('scroll', updateArrows);
+    globalThis.addEventListener('resize', updateArrows);
+    return () => {
+      el.removeEventListener('scroll', updateArrows);
+      globalThis.removeEventListener('resize', updateArrows);
+    };
+  // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [dependency]);
+
+  const scroll = (dir: 'left' | 'right') => {
+    scrollRef.current?.scrollBy({ left: dir === 'left' ? -CARD_WIDTH * 2 : CARD_WIDTH * 2, behavior: 'smooth' });
+  };
+
+  return { scrollRef, canScrollLeft, canScrollRight, scroll };
+}
